Move page types to page.go
